First set of changes involving migration towards non-malloc functions/vars, and implementation of consistent linked lists; not compilable yet

+++ ircbotd/sighand.c
@@ -15,8 +15,11 @@ void sighandle_parent(int sig)
 	if (globals.size > 0)
 	{
 		int i;
+		int asig = SIGUSR1;
+		if (sig == SIGSEGV)
+			asig = SIGSEGV;
 		for (i = 0; i < globals.size; i++)
-			if (globals.config[i].pid != 0) kill(globals.config[i].pid, SIGUSR1);
+			if (globals.config[i].pid != 0) kill(globals.config[i].pid, asig);
 		kill(globals.lib_pid, SIGUSR1);
 	}
 	free_ninfo(globals.config, globals.size);
@@ -42,13 +45,18 @@ void sighandle_child(int sig)
 		alarm(PING_DELAY);
 	}
 	#endif
-	else if (sig != SIGUSR1)
+	else if (sig != SIGUSR1 && sig != SIGSEGV)
 		kill(globals.parent_pid, sig);
 	else
 	{
+		char * quitmsg = "QUIT :Terminated by signal\n";
+		if (sig == SIGSEGV) quitmsg = "QUIT :Segmentation fault!\n";
 		globals._run = 0;
 		if (globals.confPtr->enabled)
-			write_data(globals.confPtr->sockfd, "QUIT :Terminated by signal\n");
+		{
+			usleep(UDELAY);
+			write_data(globals.confPtr->sockfd, quitmsg);
+		}
 		else
 			close(globals.confPtr->sockfd);
 	}
@@ -57,11 +65,8 @@ void sighandle_child(int sig)
 void sighandle_lib(int sig)
 {
 	if (sig == SIGSEGV)
-	{
 		irc_printf(IRCERR, "Segmentation fault!\n");
-		kill(globals.parent_pid, SIGTERM);
-	}
-	else if (sig != SIGUSR1)
+	if (sig != SIGUSR1)
 		kill(globals.parent_pid, sig);
 	else
 		globals._run = 0;
@@ -86,13 +91,7 @@ int set_signals(int type)
 		case _PARENT:
 		newaction.sa_handler = sighandle_parent;
 		newaction.sa_flags = newaction.sa_flags | SA_RESETHAND;
-		if ((sigaction(SIGABRT, &newaction, NULL) + sigaction(SIGTERM, &newaction, NULL) + sigaction(SIGINT, &newaction, NULL)) < 0)
-		{
-			irc_printf(IRCERR, "Error setting signal handlers: %s\n", strerror(errno));
-			return -1;
-		}
-		newaction.sa_flags = newaction.sa_flags;
-		if ((sigaction(SIGUSR1, &newaction, NULL)) < 0)
+		if ((sigaction(SIGUSR1, &newaction, NULL) + sigaction(SIGSEGV, &newaction, NULL) + sigaction(SIGABRT, &newaction, NULL) + sigaction(SIGTERM, &newaction, NULL) + sigaction(SIGINT, &newaction, NULL)) < 0)
 		{
 			irc_printf(IRCERR, "Error setting signal handlers: %s\n", strerror(errno));
 			return -1;
@@ -101,7 +100,7 @@ int set_signals(int type)
 		case _CHILD:
 		newaction.sa_handler = sighandle_child;
 		newaction.sa_flags = newaction.sa_flags | SA_RESETHAND;
-		if ((sigaction(SIGUSR1, &newaction, NULL) + sigaction(SIGABRT, &newaction, NULL) + sigaction(SIGTERM, &newaction, NULL) + sigaction(SIGINT, &newaction, NULL)) < 0)
+		if ((sigaction(SIGSEGV, &newaction, NULL) + sigaction(SIGUSR1, &newaction, NULL) + sigaction(SIGABRT, &newaction, NULL) + sigaction(SIGTERM, &newaction, NULL) + sigaction(SIGINT, &newaction, NULL)) < 0)
 		{
 			irc_printf(IRCERR, "Error setting signal handlers: %s\n", strerror(errno));
 			return -1;
